sustenance/requirements.txt
Django==2.0.2
django-webpack-loader==0.5.0
djangorestframework==3.7.7
wheel==0.30.0
mysqlclient==1.3.12